requirements.txt
agent-py
kseniaWebsocketLibrary
python-dotenv
async-timeout